tests/wpt/css-tests/css-flexbox-1_dev/html/flexbox_computedstyle_flex-flow-wrap.htm

<!DOCTYPE html>
<html><head><title>flexbox | computed style | flex-flow: wrap</title>
<link href="http://opera.com" rel="author" title="Opera Software">
<link href="http://www.w3.org/TR/css-flexbox-1/#propdef-flex-flow" rel="help">
<link href="http://www.w3.org/TR/css-flexbox-1/#valdef-flex-wrap-wrap" rel="help">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta content="dom" name="flags">
<style>
body {
	color: red;
	flex-flow: wrap;
}
</style>
</head><body><div id="log"></div>
<script>
test(function() {
    var body = document.body;

    assert_equals(getComputedStyle(body).getPropertyValue("flex-direction"),
	          "row");
    assert_equals(getComputedStyle(body).getPropertyValue("flex-wrap"),
	          "wrap");
});
</script>
</body></html>
